reorganize event code into separate files
